+++ src/routes/mod.rs
@@ -1,6 +1,6 @@
 mod challenge;
 mod authorize;
-mod credentials;
+mod authentication;
 mod did;
 mod endpoint;
 mod refresh;
@@ -10,7 +10,7 @@ mod well_known_did_config;
 pub mod error;
 pub use challenge::*;
 pub use authorize::*;
-pub use credentials::*;
+pub use authentication::*;
 pub use did::*;
 pub use endpoint::*;
 pub use refresh::*;
